requirements.txt
PySide6>=6.6
pyserial>=3.5
obsws-python>=1.7
pyautogui>=0.9
